pyproject.toml
[project]
name = "anaisp-ae-3-5"
version = "0.1.0"
description = "Yoga agent"
authors = [
    {name = "sianaisp",email = "anais.plancoulaine@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "streamlit (>=1.49.0,<2.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "beautifulsoup4 (>=4.13.5,<5.0.0)",
    "langchain (>=0.3.27,<0.4.0)",
    "langchain-openai (>=0.3.32,<0.4.0)",
    "langgraph (>=0.6.6,<0.7.0)",
    "tiktoken (>=0.11.0,<0.12.0)",
    "black (>=25.1.0,<26.0.0)"
]

[tool.poetry]
packages = [{include = "anaisp_ae"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
